+++ QCloudAPI_SDK/QCloudAPI_SDK/Module/Athena.cs
@@ -1,0 +1,10 @@
+namespace QCloudAPI_SDK.Module
+{
+    public class Athena : Base
+    {
+        public Athena()
+        {
+            serverHost = "athena.api.qcloud.com";
+        }
+    }
+}